kernel_cmdline: Refactor parsing to improve quote and whitespace handling
This moves the bulk of the parsing logic into `Parameter`.  Its
`parse` method returns a 2-tuple, its items being:

1.  An `Option<Parameter>` which contains the parsed `Parameter`.
This will be `None` if the provided input is empty or contains only
whitespace.

2.  A byte slice of the rest of the input which was not consumed by
the yielded `Parameter`.

This also adds new tests for some cases that would fail under the
previous parser implementation.

Signed-off-by: John Eckersberg <jeckersb@redhat.com>