kernel_cmdline: Update add_or_modify to better describe action taken
In some cases it will be important to differentiate whether we added a
net-new parameter vs. modifying an existing parameter.

Motivated by trying to update bootc_kargs to use kernel_cmdline.  We
want to parse the booted kargs, and then merge with any args defined
in kargs.d.  But it should error if a value in kargs.d already exists
in the booted system with a different value.  The previous API did not
provide enough visibility to make this determiniation.

Signed-off-by: John Eckersberg <jeckersb@redhat.com>