kernel_cmdline: Standardize Deref/AsRef impls across types
Implement generic AsRef in terms of Deref as suggested in:

https://doc.rust-lang.org/std/convert/trait.AsRef.html#generic-implementations

I'm not 100% sold on this since in some case (like the change to the
tests here) you have to end up doing `&*` coercion to get the right
type, but this at least makes everything consistent.

Would be a bit nicer when/if str_as_str[1] stabilizes, at least for
the most common UTF-8 case.

[1] https://github.com/rust-lang/rust/issues/130366

Signed-off-by: John Eckersberg <jeckersb@redhat.com>