Justfile: Bring back build-sealed target
Except now this wraps the "variant" usage of `just build`.  For one,
this is a convenient little piece of shorthand.  But perhaps more
importantly, we have this referenced in the docs.  Although I could go
change the docs to reflect the new usage, I think it just reads nicer
in the docs for it to remain `just build-sealed`.

Signed-off-by: John Eckersberg <jeckersb@redhat.com>